Add history blog
Creating a react page [git-log-main](/git-log-main) for visualising the
git log as a timeline.

As the commits are authored using markdown, the library `react-markdown`
is used to render
each timeline element.

The component used for visualizing the timeline is
`rect-vertical-timeline`, the github commits API is used to get the
history of the branch
https://stephane-monnot.github.io/react-vertical-timeline